add ScctPlugin to com.sqality.scct.sbt package and deprecate ScctPlugin in default package
